feat: Add support for mail configuration in notification settings
This commit adds support for configuring email notifications in the notification settings. It includes updates to the language files to add translations for the new "Mail Config" tab, as well as changes to the layout and components to display and manage the mail configuration options. This enhancement improves the flexibility and usability of the notification settings.